Deleted First exercises (for ex. 2)
Apparently it creates the folder again and just move the file to the specified folder. Pretty... strange... Strange because that way you can have multiple folders with the same name and description which is very inconvenient. Hence a lot of problems/issues are possible to happen. Need to remember this 'con'.